12 de mayo ajustes roles, y pedidos
