+++ PortalClientesRotoplas/src/main/java/com/alliax/portalclientes/locale/Labels.properties
@@ -451,3 +451,9 @@ contactanos=Contáctanos
 listaClientes=Lista de Clientes
 carteraVencida=Cartera Vencida
 facturasPorVencer=Facturas por Vencer
+
+#cancelarPedido
+confirmacion=Confirmación
+cancelarPedidosConfirmacion=¿ Desea cancelar el pedido ?
+pedidoNoCancelado = Pedido no ha sido Cancelado 
+pedidoCancelado = Pedido ha sido Cancelado
