+++ PortalClientesRotoplas/src/main/java/com/alliax/portalclientes/locale/Labels_es.properties
@@ -491,10 +491,7 @@ btnCancelarPedido = Cancelar Pedido
 btnCopiarPedido = Copiar Pedido
 
 #cancelarPedido
-#cancelarPedidos
 confirmacion=Confirmación
 cancelarPedidosConfirmacion=¿ Desea cancelar el pedido ?
-
-
-pedidoNoCancelado = Pedido no ha podido ser Cancelado 
+pedidoNoCancelado = Pedido no ha sido Cancelado 
 pedidoCancelado = Pedido ha sido Cancelado
