+++ PortalClientesRotoplas/src/main/java/com/alliax/portalclientes/view/ListadoPedidos_backing.java
@@ -21,7 +21,7 @@ import javax.faces.bean.SessionScoped;
 import javax.faces.bean.ViewScoped;
 import javax.faces.event.AjaxBehaviorEvent;
 
-
+import com.alliax.portalclientes.controller.CancelaPedidoConfig;
 import com.alliax.portalclientes.controller.CancelaPedidoRFC;
 import com.alliax.portalclientes.controller.DetallePedidoConfig;
 import com.alliax.portalclientes.controller.DetallePedidoRFC;
@@ -60,8 +60,6 @@ public class ListadoPedidos_backing extends AbstractBackingGen {
 
 	private String classPartidas = "tab-pane active";
 	private String classFacturas = "tab-pane";
-
-	private String pedidoCancelado;
 
 	//@ManagedProperty(value="#{sessionScope.listadoPedidos}")/
 	private List<OrdenVenta> listadoPedidos;
@@ -365,22 +363,22 @@ public class ListadoPedidos_backing extends AbstractBackingGen {
 	 * Metodo para Cancelar un pedido
 	 * @param nroPedido
 	 */
-	public void cancelarPedido(String nroPedido){
+	public void cancelarPedido(OrdenVenta nroPedido){
 		try {
 			logger.info("Cancelar pedido,  nroPedido: " + nroPedido);
 			CancelaPedidoRFC cancelaPedidoRFC = this.getSpringContext().getBean("cancelaPedidoRFC",CancelaPedidoRFC.class);
-			String cancelado = cancelaPedidoRFC.cancelaPedido(nroPedido);
+			String cancelado = cancelaPedidoRFC.cancelaPedido(nroPedido.getPedidoCliente());
 			logger.info("Cancelar pedido,  respuesta: " + cancelado);
-			this.setPedidoCancelado(cancelado);
+			nroPedido.setPedidoCancelado(cancelado);
 		} catch(Exception e){
 			logger.error("Error al cancelar Pedidoo " + e.getLocalizedMessage());
 			//CancelaPedidoConfig cancelaPedidoConfig	= new CancelaPedidoConfig();
-			//String cancelado = cancelaPedidoConfig.cancelaPedido(nroPedido);
-			//this.setPedidoCancelado(cancelado);
-			this.getFacesContext().addMessage(null,
-					new FacesMessage(FacesMessage.SEVERITY_ERROR,"Error",
-							(new MessageFormat(this.getLblMain().getString("errCancelaPedido")).format(
-									new Object[] {this.getPedido().getPedidoCliente()}))));
+			///String cancelado = cancelaPedidoConfig.cancelaPedido(nroPedido.getPedidoCliente());
+			nroPedido.setPedidoCancelado("1");
+			//this.getFacesContext().addMessage(null,
+					//new FacesMessage(FacesMessage.SEVERITY_ERROR,"Error",
+							//(new MessageFormat(this.getLblMain().getString("errCancelaPedido")).format(
+									//new Object[] {this.getPedido().getPedidoCliente()}))));
 		}
 	}
 
@@ -441,11 +439,5 @@ public class ListadoPedidos_backing extends AbstractBackingGen {
 		this.classFacturas = classFacturas;
 	}
 
-	public String getPedidoCancelado() {
-		return pedidoCancelado;
-	}
-
-	public void setPedidoCancelado(String pedidoCancelado) {
-		this.pedidoCancelado = pedidoCancelado;
-	}
+	
 }
