+++ PortalClientesRotoplas/src/main/webapp/WEB-INF/spring/security-context.xml
@@ -33,7 +33,13 @@
 		                                               ROLE_EJECUTIVO_MEMBER,ROLE_EJECUTIVO_MEMBER_AR,ROLE_EJECUTIVO_MEMBER_PE,ROLE_EJECUTIVO_MEMBER_CA,ROLE_EJECUTIVO_MEMBER_BR" />
 		<intercept-url pattern="/resources/files/**" access="ROLE_ANONYMOUS,ROLE_CLIENT_MEMBER,ROLE_ADMIN_MEMBER,ROLE_VENTAS_MEMBER,ROLE_EJECUTIVO_MEMBER,ROLE_ADMIN_MEMBER_AR,ROLE_VENTAS_MEMBER_AR,ROLE_EJECUTIVO_MEMBER_AR,ROLE_ADMIN_MEMBER_PE,ROLE_VENTAS_MEMBER_PE,ROLE_EJECUTIVO_MEMBER_PE,ROLE_ADMIN_MEMBER_CA,ROLE_VENTAS_MEMBER_CA,ROLE_EJECUTIVO_MEMBER_CA,ROLE_ADMIN_MEMBER_BR,ROLE_VENTAS_MEMBER_BR,ROLE_EJECUTIVO_MEMBER_BR" />
 		<intercept-url pattern="/resources/filesAdm/**" access="ROLE_ADMIN_MEMBER,ROLE_ADMIN_MEMBER_AR,ROLE_ADMIN_MEMBER_PE,ROLE_ADMIN_MEMBER_CA,ROLE_ADMIN_MEMBER_BR" />
-					
+
+		<intercept-url pattern="/pedidos/crearPedido.xhtml" access="ROLE_PEDIDOS,ROLE_ADMIN_MEMBER,ROLE_ADMIN_MEMBER_AR,ROLE_ADMIN_MEMBER_PE,ROLE_ADMIN_MEMBER_CA,ROLE_ADMIN_MEMBER_BR" />
+		<intercept-url pattern="/pedidos/listado.xhtml" access="ROLE_PEDIDOS,ROLE_SEGUIMIENTO_PEDIDOS,ROLE_ADMIN_MEMBER,ROLE_ADMIN_MEMBER_AR,ROLE_ADMIN_MEMBER_PE,ROLE_ADMIN_MEMBER_CA,ROLE_ADMIN_MEMBER_BR" />
+		<intercept-url pattern="/cotizaciones.xhtml" access="ROLE_PEDIDOS,ROLE_ADMIN_MEMBER,ROLE_ADMIN_MEMBER_AR,ROLE_ADMIN_MEMBER_PE,ROLE_ADMIN_MEMBER_CA,ROLE_ADMIN_MEMBER_BR" />
+		<intercept-url pattern="/consultaExistencia.xhtml" access="ROLE_PEDIDOS,ROLE_SEGUIMIENTO_PEDIDOS,ROLE_ADMIN_MEMBER,ROLE_ADMIN_MEMBER_AR,ROLE_ADMIN_MEMBER_PE,ROLE_ADMIN_MEMBER_CA,ROLE_ADMIN_MEMBER_BR" />
+		<intercept-url pattern="/estadoCuenta.xhtml" access="ROLE_ESTADO_CUENTA,ROLE_ADMIN_MEMBER,ROLE_ADMIN_MEMBER_AR,ROLE_ADMIN_MEMBER_PE,ROLE_ADMIN_MEMBER_CA,ROLE_ADMIN_MEMBER_BR" />
+
 		<intercept-url pattern="/recuperaPassword.xhtml" access="ROLE_ANONYMOUS" />
 		<intercept-url pattern="/indexPwdChange.xhtml" access="ROLE_PWDCHANGE_MEMBER" />
